backend/fastapi/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.2
certifi==2025.8.3
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.25.0
pytest-cov==6.3.0

# Security scanning tools
bandit==1.7.10
safety==3.2.11
pip-audit==2.7.3
detect-secrets==1.5.0

# Code quality and linting tools
flake8==7.1.1
pylint==3.3.2
mypy==1.13.0
black==24.10.0
isort==5.13.2

# Type stubs
types-passlib==1.7.7.20240819
types-python-jose==3.3.4.20240106
